+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added more `PartialEq` implementations for `String` and `&&str` comparisons.
+
 ## 0.3.0 - 2023-04-08
 
 ### Added